Address some bugs, refactor some FE
