+++ packages/backend/src/utils/getNumberWithinRange.ts
@@ -1,0 +1,3 @@
+export function getNumberWithinRange(baseNumber: number, min: number, max: number): number {
+    return Math.max(Math.min(baseNumber, max), min);
+}
