+++ packages/frontend/src/components/portfolio/helperComponents/marketStatus.module.scss
