+++ packages/frontend/src/components/stocks/helperComponents/customTapValueIndicator.module.scss
