+++ packages/frontend/src/components/stocks/helperComponents/stockChart.module.scss
