+++ packages/frontend/src/components/stocks/helperComponents/stocksDialog.module.scss
