--- packages/frontend/src/selectors/selectUserOwnedStock.ts
@@ -1,16 +1,0 @@
-import { IOwnedStock, IStockWithDollarValue } from "@stochastic-exchange/api";
-import { createSelector } from "reselect";
-import { IStoreState } from "../store/state";
-
-export function selectUserOwnedStock(stock: IStockWithDollarValue | undefined) {
-    return createSelector(
-        (state: IStoreState) => state.account.ownedStocks,
-        (ownedStocks: IOwnedStock[] | undefined): IOwnedStock | undefined => {
-            if (stock === undefined) {
-                return undefined;
-            }
-
-            return ownedStocks?.find(s => s.stock === stock.id);
-        },
-    );
-}
